Pair sum in array
Given a random integer array A and a number x. Find and print the pair of elements in the array which sum to x.

Array A can contain duplicate elements. While printing a pair, print the smaller element first. That is if a valid pair is (6, 5) print "5 6". There is no constraint that out of 5 pairs which have to be printed in 1st line. You can print pairs in any order, just be careful about the order of elements in a pair.